requirements.txt
aenum==3.1.15
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
klaviyo-api==14.0.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.3.0
six==1.16.0
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.2.3
